requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
